docs
Test assignment for the junior developer position. It was necessary to create api, through django rest framework to use built-in technologies to implement search for cats by name, and the same api should support sorting by a couple of fields at once, also pagination. 

A separate condition was to run detailed tests using built-in technology unittest, which could cover api in detail and identify all possible errors in views, models, serializer, pagination and filters.


In order to make this api work for you, you will have to install in a virtualized environment:

'django',
'djangorestframework',
'django-filter'.